Added modules and made them work
