+++ homework10/countries_and_capitals/graph.h
@@ -8,6 +8,7 @@ void addVertex(Graph *graph, const int vertex);
 void addEdge(Graph *graph, const int vertexA, const int vertexB, const int newLength);
 std::vector<int> adjacent(Graph *graph, const int vertex);
 bool isAdjacent(Graph *graph, const int vertexA, const int vertexB);
-int belongs(Graph *graph, const int vertex);
-void assign(Graph *graph, const int vertex, const int target);
-
+bool belongs(Graph *graph, const int vertex);
+void assign(Graph *graph, const int vertex);
+int edgeLength(Graph *graph, const int vertexA, const int vertexB);
+void printGraph(Graph *graph);
